Internal comments
Comentarios internos en el código.

+++ Laravel/viver-cellers/database/migrations/2024_05_06_075422_create_products_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->string('image_path')->nullable();
             $table->string('name');
             $table->string('winemaking');
-            $table->integer('grade_alcohol');
+            $table->decimal('grade_alcohol', 8, 2);
             $table->integer('stock');
             $table->decimal('price', 8, 2);
             $table->integer('iva');
